+++ kipper/core/src/compiler/semantics/language/expressions.ts
@@ -50,11 +50,12 @@ import {
 } from "../const";
 import type { TargetASTNodeCodeGenerator } from "../../translation";
 import type { TargetASTNodeSemanticAnalyser } from "../target-semantic-analyser";
-import { CompoundStatement } from "./statements";
 import { CompilableASTNode } from "../../parser";
-import { ScopeVariableDeclaration } from "../../scope-declaration";
+import { ScopeDeclaration, ScopeVariableDeclaration } from "../../scope-declaration";
 import { KipperNotImplementedError, UnableToDetermineMetadataError } from "../../../errors";
 import { TerminalNode } from "antlr4ts/tree";
+import { getConversionFunctionIdentifier } from "../../../utils";
+import { kipperInternalBuiltIns } from "../../runtime-built-ins";
 
 /**
  * Every antlr4 expression ctx type
@@ -561,7 +562,7 @@ export class IdentifierPrimaryExpression extends Expression<IdentifierPrimaryExp
 		// Make sure the referenced variable even exists!
 		const ref = this.programCtx
 			.semanticCheck(this)
-			.getExistingReference(identifier, this.scope instanceof CompoundStatement ? this.scope : undefined);
+			.getExistingReference(identifier, "localScope" in this.scopeCtx ? this.scopeCtx : undefined);
 
 		// Evaluate the type by attempting to fetch it from the global
 		const evaluateType: () => KipperType = () => {
@@ -571,10 +572,15 @@ export class IdentifierPrimaryExpression extends Expression<IdentifierPrimaryExp
 				return "func";
 			}
 		};
+
 		this.semanticData = {
 			evaluatedType: evaluateType(),
 			identifier: identifier,
 		};
+
+		if (!(ref instanceof ScopeDeclaration)) {
+			this.programCtx.addBuiltInReference(this, ref);
+		}
 	}
 
 	/**
@@ -1389,6 +1395,13 @@ export class CastOrConvertExpression extends Expression<CastOrConvertExpressionS
 			type: type,
 			exp: exp,
 		};
+
+		// Add internal reference to the program ctx.
+		const expType = (<Expression<ExpressionSemantics>>exp).getSemanticData().evaluatedType;
+		const internalIdentifier = getConversionFunctionIdentifier(expType, type);
+		if (internalIdentifier in kipperInternalBuiltIns) {
+			this.programCtx.addInternalReference(this, kipperInternalBuiltIns[internalIdentifier]);
+		}
 	}
 
 	/**