+++ kipper/core/src/utils.ts
@@ -2,7 +2,6 @@ import { Interval } from "antlr4ts/misc/Interval";
 import { ParserRuleContext, Token } from "antlr4ts";
 import { ParseTree } from "antlr4ts/tree";
 import { CharStream } from "antlr4ts/CharStream";
-import { CompilableASTNode, CompoundStatement, compilableNodeParent, KipperScope, RootASTNode } from "./compiler";
 
 /**
  * Returns the token source for the passed {@link antlrCtx} instance.
@@ -54,25 +53,6 @@ export function getTokenSource(inputStream: CharStream, token: Token) {
  */
 export function getParseTreeSource(inputStream: CharStream, parseTree: ParseTree) {
 	return inputStream.getText(parseTree.sourceInterval);
-}
-
-/**
- * Determine the scope of the token.
- * @param ctx The token ctx.
- * @since 0.4.0
- */
-export function determineScope(ctx: CompilableASTNode<any>): KipperScope {
-	// Determine type by going up the parent structure, until a compound statement is hit or the root file parse
-	// token, which represents the entire programCtx.
-	let parent: compilableNodeParent = ctx.parent;
-	while (!(parent instanceof RootASTNode) && !(parent instanceof CompoundStatement)) {
-		parent = parent.parent;
-	}
-
-	if (parent instanceof RootASTNode) {
-		return ctx.programCtx;
-	}
-	return parent;
 }
 
 /**