+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#134](https://github.com/Luna-Klatzer/Kipper/issues/134))
 - Implemented reserved identifier checking, which ensures that no declarations overwrite/interfere with an internal
   identifier or reserved identifier/keyword. ([#153](https://github.com/Luna-Klatzer/Kipper/issues/153))
+- Implemented tree-shaking for internal and built-in functions using the new class `KipperOptimiser`, which removes
+  any function definitions that are not used ([#159](https://github.com/Luna-Klatzer/Kipper/issues/159)).
 - New field `KipperCompileTarget.builtInGenerator`, which will store the built-in generator for each target.
 - New classes and interfaces:
   - `KipperTargetBuiltInGenerator`, which updates the behaviour for generating built-in functions.
@@ -38,9 +40,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     keywords and other internal logic.
   - `KipperSemanticErrorHandler`, which implements a default abstract error handler for semantic errors. This is
     used by `KipperTargetSemanticAnalyser` and `KipperAsserter`.
+  - `KipperOptimiser`, which handles code optimisation for a Kipper program.
 - New functions:
   - `KipperSemanticChecker.validConversion()`, which checks whether a type conversion is valid and implemented by
     Kipper.
+  - `KipperOptimiser.optimise()`, which performs optimisation on an abstract syntax tree.
+  - `KipperProgramContext.getBuiltInFunction()`, which searches for a built-in function based on a passed
+    identifier.
+  - `KipperProgramContext.optimise()`, which performs code optimisations for the local abstract syntax tree.
 - New errors:
   - `InvalidConversionError`, which is thrown when an invalid or unimplemented conversion is performed in a Kipper
     program.
@@ -50,6 +57,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Kipper meta type `type`, which represents the type of a Kipper type.
   - `kipperSupportedConversions`, which is an array containing multiple tuples representing allowed conversions in
     Kipper.
+- New `CompileConfig` option `optimisationOptions`, which contains the configuration for the `KipperOptimiser`.
+- Added new flags `-b/--[no-]optimise-builtins` and `-i/--[no-]optimise-internals` to the `@kipper/cli` for enabling
+  internal and built-in functions optimisation.
 
 ### Changed
 
@@ -68,12 +78,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed:
   - `builtIns` to `kipperRuntimeBuiltIns`.
   - `semantic-analyser.ts` to `target-semantic-analyser.ts`.
-  - `ParserASTNode.ensureTokenChildrenExist` to `getTokenChildren`.
-  - `ParserASTNode.ensureSemanticDataExists` to `getSemanticData`.
-  - `KipperError.setMetadata` to `KipperError.setTracebackData()`.
+  - `ParserASTNode.ensureTokenChildrenExist()` to `getTokenChildren()`.
+  - `ParserASTNode.ensureSemanticDataExists()` to `getSemanticData()`.
+  - `KipperError.setMetadata()` to `KipperError.setTracebackData()`.
   - `KipperAsserter` to `KipperSemanticsAsserter`.
   - `TargetTokenCodeGenerator` to `TargetASTNodeCodeGenerator`.
   - `TargetTokenSemanticAnalyser` to `TargetASTNodeSemanticAnalyser`.
+  - `CompilerEvaluatedOptions` to `EvaluatedCompileOptions`.
+  - `KipperProgramContext.processedParseTree` to `abstractSyntaxTree`.
+  - `KipperProgramContext.builtInGlobals` to `builtIns`.
+  - `CompileConfig.globals` to `builtIns`.
+  - `CompileConfig.extendGlobals` to `extendBuiltIns`.
 - Optimised and simplified Kipper code generation in `KipperCompileResult.write()`.
 - Updated `@kipper/core` code base structure:
   - `/parser/` now contains these new files:
@@ -83,6 +98,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `/semantics/language/` which contains the language specific AST Node classes that implement the semantics for the
     expressions, definitions and statements in Kipper.
   - `/semantics/processor/` which is the module containing the Semantic analyser and Type checker.
+- Updated local and global scope handling by introducing three new classes: `Scope`, `GlobalScope` and `LocalScope`.
+  These classes now handle local variables and functions and implement a standard interface for handling declarations
+  and definitions.
+- Updated `@kipper/cli` flag names from camelCase to dash-case.
 
 ### Removed
 
@@ -98,6 +117,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for multi strings seperated by a whitespace (e.g. `"1" "2"` is counted as a single string `"12"`). This
   may be added back later, but for now it will be removed from the Kipper language.
 - Error `InvalidOverwriteError`, as all errors it represented are now subclasses of `IdentifierError`.
+- Unneeded functions from `KipperProgramContext`, as they were replaced by the new scope handling classes:
+  - `getGlobalFunction()`
+  - `getGlobalIdentifier()`
+  - `getGlobalVariable()`
+  - `addGlobalVariable()`
+- Unneeded function `determineScope()`, as the scope handling was moved to `CompilableASTNode.scope`.
 
 ## [0.7.0] - 2022-05-22